+++ Horizontes_Urbanos/src/main/java/Modelo/jpa_controllers/AsesorJpaController.java
@@ -163,7 +163,7 @@ public class AsesorJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                int id = asesor.getCedula();
+                Long id = asesor.getCedula();
                 if (findAsesor(id) == null) {
                     throw new NonexistentEntityException("The asesor with id " + id + " no longer exists.");
                 }
@@ -176,7 +176,7 @@ public class AsesorJpaController implements Serializable {
         }
     }
 
-    public void destroy(int id) throws IllegalOrphanException, NonexistentEntityException {
+    public void destroy(Long id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -239,7 +239,7 @@ public class AsesorJpaController implements Serializable {
         }
     }
 
-    public Asesor findAsesor(int id) {
+    public Asesor findAsesor(Long id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Asesor.class, id);
